+++ CopilotKit/packages/sdk-js/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @copilotkit/sdk-js
 
+## 1.6.0-next.9
+
+### Patch Changes
+
+- @copilotkit/shared@1.6.0-next.9
+
+## 1.6.0-next.8
+
+### Patch Changes
+
+- @copilotkit/shared@1.6.0-next.8
+
+## 1.6.0-next.7
+
+### Patch Changes
+
+- @copilotkit/shared@1.6.0-next.7
+
 ## 1.6.0-next.6
 
 ### Patch Changes