+++ CopilotKit/packages/shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @copilotkit/shared
 
+## 1.6.0-next.9
+
+## 1.6.0-next.8
+
+## 1.6.0-next.7
+
 ## 1.6.0-next.6
 
 ## 1.6.0-next.5